Update build system to CMake
Add CMake as the official build system!

Currently, the project does not compile, but it is now possible to
actually build the project via CMake. Run ./build.sh in order to speed
up the process by generating the build system and then actually building
the project.

Note that this project is using C++20, and that requires a compiler
that supports it. The oldest version of GCC that supports all C++20
features is gcc-12, with almost all features being supported in gcc-11.
See https://gcc.gnu.org/projects/cxx-status.html#cxx20 for more
information