Add parsing function to tests
Parse the disassembled output (address in hex\topcode\n) so that tests
can be written in a much easier way. At this point, future unit tests
can simply setup the input that they want to decode, and then inspect
the list of addresses & opcodes & make sure that they are expected

Additionally, introduce a typedef for the address type. A couple of
times while I was programming this commit, I mistakenly used byte as
the type to represent addresses, but that's incorrect, so the type has
been centralized.