STA implementation, unit test improvements
Implementation of the STA opcode (0x32).

The STA opcode requires 3 bytes of data: the opcode, a low byte, and a
high byte. The low and high byte, when concatenated, described the value
that should be stored in the accumulator.

Also, streamline unit testing, so that parsing disassembled output
can be achieved in two lines -- creating the desired input, and then
parsing the output.

Introduce the opcode enum for easier testing & matching in the main
decode switch statement. I can see this being confusing since there
are kind of two versions of the opcode that are relevant in the decoding
process: the raw value of it (a byte value), and the textual
representation (a std::string). Sometimes we deal with one, sometimes
we deal with the other, but the opcode enum is only meant for the former