LDA implementation, STA refactor
The LDA instruction (0x3a, 0b00111010) loads the contents of the
specified memory loction insto the accumulator. Note that the 8080 is a
little endian machine, so the following two bytes are the low byte & the
high byte, respectively.

Also, refactor some existing STA code to make dealing with multi-byte
opcodes slightly easier.